Convert crypt-gpgme to use BUFFER for tempfiles.
Use the pool where possible, but when the function calls other
long-running functions (such as mutt_body_handler() or
mutt_do_pager()), allocate the buffer.

smime_gpgme_decrypt_mime() was not freeing the "cur" BODY generated on
error.  Add a call to mutt_free_body(cur) in the error handling.