Ensure APP_PGP flag is enabled for autocrypt.
There was no code that explicitly enabled APP_PGP mode when either the
recommendation engine or autocrypt menu were invoked.  This would
cause the actual encryption to fail when sending.

This was hidden by the $crypt_opportunistic_encrypt config, which does
ensure it gets turned on.  I can only guess that this variable is
popular with the autocrypt early adopters too.

Also explicity enable APP_PGP for $autocrypt_reply.  We can't rely on
$crypt_autopgp to turn this on.