Fix copyright header in mutt_zstrm.[ch]
In <https://gitlab.com/muttmua/mutt/merge_requests/60#note_252772011>
Fabian noticed that he copied the header file as a template but forgot
to adjust the copyright line at the top to himself.

Also, adjust the formats in both to match the rest of mutt, so the
copyright updater script will work properly.