Fix new option order in the manual.
I actually intended to push commit abfaa0e0 to a remote branch, but
accidentally ended up pushing it to stable. :-O

Anyway, I think the option is okay, but may have to tweek it a bit.
This fixes a mistake where it was placed out of order in init.h.