Convert makedoc.c to makedoc.pl.
This removes the build-time dependency on a compiled program, which
doesn't work for cross-compilation.

My perl is pretty rusty, and the result isn't super pretty.  However,
since perl is already a build dependency, it made the most sense to
use that.  The output of Muttrc, muttrc.man, and manual.xml match.

There might be a few bugs, so I'm going to leave makedoc.c in for now
as a reference.