Add move-up and move-down functions to compose menu.
This adds the ability to rearrange the order of the attachments,
including swapping the first attachment if desired.

The compose menu doesn't support nested attachments, but since data
structures and code are shared with recvattach, the new functions use
the v2r and try to behave sensibly with respect to parent pointer
adjustments.

However, the functions do restrict movement to be between siblings.
Adding code to support cross-tree movement would be more complicated
for no current purpose.