Fix ~v tagging when switching to non-threaded sorting.
Place num_hidden in all the collapsed thread messages, rather than
just the visible ones (i.e. the root headers).

This is a tweak to commit 0f8a079d, where I discovered the behavior
inconsistent with the documentation.  I'm making this change in master
due to the (tiny) behavior change.