Ensure format string dest buffer is cleared for callback function.
There is some variation in Mutt as to the behavior of an undefined
expando.  Some functions pass the op through, some output nothing.

Unfortunately, a few callback functions also don't ensure the dest
buffer is cleared if no expandos match the passed in op.  This is
mostly my fault - I think I added a few using addrbook as a template,
which was unfortunately missing the default clear switch.

Rather than adding it to the places a default is missing, just ensure
it's cleared in mutt_FormatString() itself, which will prevent future
mistakes too.