Fix $recall prompting.
The background-edit added a flag for the <mail> operation, so the
previous test for !flags now always fails.

I don't see a more elegant fix, other than changing it to match
SENDBACKGROUNDEDIT, plus the existing test for sctx->msg being NULL.

Add a comment because the flag test is confusing otherwise.