Change mutt_FormatString() data parameter to type void *.
All invocations pass a pointer parameter.  Perhaps at one point this
was like the init.h combined usage.  But as with that, using an
unsigned long is incorrect, and we've been lucky to get away with it.
Since we don't need a union, just change it to void * to properly size
the parameter.