Add $folder to send scope.
It's hard to know where to draw the line on scope variables, but
$folder seems likely to change across "account switches" and will
affect Fcc pretty_mailbox and expansion.  The user can also manually
change Fcc from the compose menu.

If spoolfile or infile complaints come in, I can add that, but I'm not
convinced those will commonly be an issue for sending emails.