Add $fcc_delimiter, to allow multiple Fcc mailboxes.
The variable defaults unset, to prevent breaking anyone's
configuration.

If set to a value, such as ",", $record and fcc-hook may list multiple
mailboxes separated by the delimiter.  Mutt will tokenize and
expand/prettify each separate mailbox.  At Fcc time, it will loop
through trying to save to each mailbox.

This patch introduces a bit more uglyness into init.c, which has to
special-process for $record in a few places.  I don't like that, but I
believe the feature is important enough to allow for the one-off
behavior hack, because we are constrained by backwards compatibility.

Add documentation notes about the resulting issues with '<' mailbox
shortcut expansion, and with fcc-save-hook processing.  It is
recommended to use fcc-hook to set multiple mailboxes in those cases
instead.