Convert mutt_parse_rc_line() to use real buffer to hold the line.
Previously, the parameter converted into a kind of read-only buffer,
pointing to the char *line passed in.  The problem is that
mutt_extract_token() backtick processing allocates and assigns a new
token->data.

As a result, buffer->destroy was added, whose sole purpose is to keep
track of the read-only token buffer being reallocated so that it can
be properly freed inside mutt_parse_rc_line().  (The char *line is
allocated and freed in source_rc()).

Remove the token parameter, and convert line to a const char *.  Copy
that into a buffer-pool buffer, so the buffer can be modified
"normally".

Create a mutt_buffer_rc_buffer() function taking the line buffer
parameter.  To make the source_rc() just a tiny bit faster, have it
directly manage and call mutt_parse_rc_buffer() itself.

Other callers likely don't need the speed, so remove their cumbersome
creation/free of that parameter.

The next commit will deal with removing buffer->destroy and fixing up
backtick handling.