Display error if no imap authenticators are available.
Without SASL enabled, and with an explicit value in
$imap_authenticators not natively implemented by Mutt, no calls to
authenticator->authenticate() were occuring.

This resulted in the default r value of -1 passing through, which did
not trigger the mutt_error() message after the loop.

Without that message, an empty index would display with the
"connecting" message still on the bottom, giving the appearance that
Mutt is hung.