Convert "non-fatal" handler errors to return 1 instead of -1.
When displaying a message, display a mutt_error() to warn of
incomplete rendering.

Fix mutt_copy_message() to check for ferror and feof errors on partial
decode too.

Clean up _mutt_append_message() to not pass a partial-decode along as
a success.

Modify the crypt handlers to return 1 if any kind of state message is
displayed.  There is some fuzzyness about what a "fatal" error is, but
for now just consider a handler error that notifies by
state_attach_puts() as "non-fatal".