olen length requirement was too conservative
When converting binary values to Base 64 encoding, we go from a 8 bit
representation to a 6 bit representation. As a result, we need extra space
to store the result compared to the input value being converted.

In the current implementation, the requirements on the size of the output
buffer are too conservative; We only have to check if the remaining space
is larger than 4 (3 input bytes get converted into 4 Base64 encoded output
bytes, plus allowing for a terminating '\0' character) rather than 10.