Support domain-literal values in email adddresses.
RFC 5322 defines the domain-literal:

   domain          =   dot-atom / domain-literal / obs-domain

   domain-literal  =   [CFWS] "[" *([FWS] dtext) [FWS] "]" [CFWS]

   dtext           =   %d33-90 /          ; Printable US-ASCII
                       %d94-126 /         ;  characters not including
                       obs-dtext          ;  "[", "]", or "\"

RFC 5321 defines possible literal values in section 4.1.3.

I think this was unnoticed (and possibly unimplemented) because it's a
rare use-case.  However, Stuart Gathman @sdgathman in the
ticket (#226) noted:

  The use case is replacing alpine with mutt for this fully
  decentralized usage:

  https://fedoramagazine.org/decentralize-common-fedora-apps-cjdns/

  The only requirements are to support IPv6 and IPv6 literals (to
  avoid use of DNS).

Adding this slightly complicates the parser, because we can't wait to
decide whether it's a domain-literal until parse_mailboxdomain(), but
we also need to accommodate the CFWS.

Break out a parse_domain() function, with partial duplication of
parse_mailboxdomain()'s logic.

Also, add parse_literal() to the outer parser, because we need to
handle both name-addr and top level addr-spec.

The RFC822Errors[] are currently unused (and broken), but add a new
error code and message, in case those get fixed up and introduced to
the code in the future.

Disable IDNA encoding for domain-literals.