Ensure idata->reopen is clear when the mailbox is closed.
Commit ab457327 accidentally introduced a performance regression in
the 1.13.0 release.  The reason is that imap_sync_mailbox() was
leaving IMAP_REOPEN_ALLOW set.  When a mailbox was closed, and another
opened, the imap_open_mailbox() would be processed with the reopen
set.

The initial connection and exec statements would trigger a
imap_cmd_finish().  That would then prematurely trigger fetching new
mail, with the newMailCount misset, which would in turn generate a
tiny uid_hash.  When the real download occurred, it would use the
existing tiny hash, throttling performance.

The regression commit removed what I thought was an unnecessary check,
but it was guarding the fetch from happening at the wrong time during
mailbox open too.

The performance issue was already addressed by commit a8337951 in
master, but I think it's good to wipe the entire state.  The removes
other possible strange behavior.

A HUGE thanks to Guilhem Moulin for his help on this issue.  Without
all his work detailing the problem, creating automated testing, git
bisecting, and logging, I would not have been able to find and fix the
problem.