Add missing doc built distfiles.
When building the 1.14.0 release tarball, I discovered a hidden
dependency problem from switching to makedoc.pl.

The manual.xml is not properly defined, and so trying to build the
muttrc.man and mutt.info file errors out.  This was previously hidden
because makedoc need to be compiled, forcing a rebuild.

This problem derives because the stamp-* files (and most of the other
documentation) are *included* in the tarball, except for muttrc.man
and mutt.texi.  Those attempt to rebuild, but manual.xml doesn't
exist.

For now, include muttrc.man and mutt.texi in the dist files.

A better solution requires some thought, but not while I'm trying to
get a release out.