Fix doc installation directories.
The conversion of makedoc.c to perl goofed up the documentation
build/install process from a tarball.  A late commit, 3c575cab, tried
to fix this, but didn't fix the problem when building with a separate
srcdir.

Previously, because makedoc didn't exist, all the documentation would
be rebuilt, even though it was distributed in the tarball.

Now the files don't need to be rebuilt, and so exist in $(srcdir)
instead of the build directory.

I'm a little bit unsure how this would all work, trying to install
from git with a separate srcdir.  I think it would fail now.  However,
I don't believe that's a common use case.

If I'm wrong, then an alternative would be to go back to rebuilding
everything.  This would be done by creating a new makedoc target, with
rule 'cp $(srcdir)/makedoc.pl makedoc', and then changing all the
other targets to depend on and invoke makedoc instead.