Separate mailbox vs buffy-incoming for the "enter_fname" functions.
Create mutt_buffer_enter_mailbox(), which turns on the MUTT_MAILBOX
flag (renamed from MUTT_EFILE).  This uses the mailbox history file
category, and turns on MUTT_SEL_FOLDER in the select_file browser.

Add a do_incoming paramter and a new separate flag MUTT_INCOMING, so
that buffy-completion can be enabled optionally for mailboxes.  It
only makes sense for the change-folder operation, so disable it in
other contexts (such as saving/copying a message, prompting for an Fcc
mailbox, or an autocrypt scan mailbox).

Change saving/copying a message to use the mailbox function, so that
it shares the history file with other mailbox operations.  It would
previously use the "file" history category.

Create mutt_buffer_enter_filenames() to replace the enter_fname()
prompt for files.

Since nothing directly uses the _mutt_buffer_enter_fname() function
anymore, rename and make it static.