Loosen message-id parser to parse ids without angle brackets.
To do this, we need to enhance the loop to skip over comments.  It
does this using the rfc822 address parse_comment() routine.

Keep parsing inside angle brackets non-strict, because of the history
of trac tickets (1116, 1935, 3090) demonstrating message-ids with
spaces inside the angle brackets, double-'@'s, or even completely
missing '@'.

However, in order to reduce the risk of extracting garbage, insist
that message-id's outside of angle brackets at least have an '@'.

Automatically add missing angle brackets so that Mutt will generate
valid references/in-reply-to headers.