Adjust message-id extraction to retry more permissively on failure.
It turns out some illegal message-id headers have neither angle
brackets nor even an '@' in them.  So adopt a different approach.

First try parsing the Message-ID header in a strict(ish) mode
requiring angle brackets.  If that returns nothing, then try again in
loose mode, which will basically grab the first non-comment token.

Change References and In-Reply-To to keep the strict(ish) mode.  We
might loose some parent references, but that won't cause much loss of
functionality in Mutt, and also helps prevents parsing in "garbage"
and passing that back in replies.

Thanks to Oswald Buddenhagen for the suggested approach.