Fix attach-file tag operation to work when quitting browser.
There was a disconnect between the compose menu and the behavior when
selecting multiple files.  The select_file() operation populates the
files and numfiles when hitting 'quit'.  The compose menu was
incorrectly looking at the fname parameter, which won't be populated
unless the user presses enter.

The enter_string() menu makes the same mistake.  Change it to behave
differently depending on whether multiple is set.

The logic needs to be made clearer by changing the api.  I will do
that in the next commit.