Remove Muttrc and manual.txt from dist tarball files.
The Muttrc needs to be generated because of docdir substitution.  The
manual.txt may traditionally be generated by a different tool than the
one on my machine (lynx), so force that to be regenerated too.

It may still be necessary to just rebuild everything, but let's see if
this will do for stable for now.