Add other headers to written Protected Headers.
In order to avoid $edit_headers, <resend-message>, or other places
setting env->date and inadvertantly generating a wrong Date header,
store the "protected headers" generated date header in the sctx, in
addition to mime_headers.

I initially removed env->date setting from mutt_parse_rfc822_line(),
thinking it a more elegant solution, but unfortunately protected
headers reading needs it to be there.  We don't currently print or
compare the headers, but might wish to in the future.

Explicitly pass in a date parameters to mutt_write_rfc822_header() to
make each caller conscious of where the date is coming from.