Consider $reply_to before $reply_self.
This is based on Derek Martin's patch in Trac ticket 2304.

I've modified it to:

* default to the value set by OPTREPLYSELF if the
  $ignore_list_reply_to short-circuit fires.

* only apply the "from==reply_to" short circuit when $reply_self would
  indicate replying to from.

* improve the "from==reply_to" short circuit to also fire when the
  display names are equal.

The discussion in the trac ticket was somewhat contentious.  This
isn't particularly an area in the code I'm eager to touch.  However,
there are three points worth noting:

1) Mail-Followup-To is considered before $reply_self, so it makes some
sense for Reply-To to also be.

2) $reply_to is a quadoption, allowing the header to be declined and
then using $reply_self if appropriate.

3) $reply_self is a boolean.  Previously, the only way to evaluate
reply_to in these cases is to set $reply_self.  Most users probably
don't want that set by default, making that a poor workaround.

If this becomes sore point, I'll revert, but I don't believe the
change will be controversial in practice.