Add first version of MuttLisp.
This is somewhat simple enhancement to muttrc processing, not a
full-blown embedded scripting language.  There are no variables,
functions, types, or abilities that compete with macros.

MuttLisp can be invoked using the "run" command.  If
$muttlisp_inline_eval is set, it can also be invoked with a bare
parenthesis expression as a command argument.

$muttlisp_inline_eval defaults unset, to avoid breaking existing
configurations, which might have bare parenthesis arguments for
regexps.