Fix redraw_motion() redraw bug.
Commit e2a28006 moved the cur_color assignment so that $arrow_cursor
could also use the value for overlays.

Unfortunately, it at the same time moved the call *after* the call to
mutt_window_move(), which would introduce a drawing bug if
menu->color() went across IMAP.

The previous commit removed overlays for $arrow_cursor, also removing
the need for cur_color to be assigned for both.  Move it back to where
it was before.