Improve w3m invocation for manual.txt generation.
This fix is also from Ambrose Li (@gniw).  The invocation makes sure
w3m won't attempt a remote connection if manual.html doesn't exist and
a proxy is configured, by using manual.txt as stdin.

It also gives a non-zero exit status in that case, and makes sure
input is processed as utf-8.