Change expand_path() to expand relative paths.
The introduction of the "cd" command can make relative paths
unreliable.  Change expand_path() to expand relative paths.

Change pretty_mailbox() to contract relative paths, but only when the
folder is outside of homedir, or when cwd is underneath the homedir.

To try and keep the sidebar from changing, use the pretty_mailbox code
to contract relative paths only.