Fix List-Post header parser to handle multiple values.
If the mailto url was in the second or greater position, beg was on
the wrong character.  This would cause url_check_scheme() to fail to
recognize the mailto.

Change the loop to skip to the '<' on subsequent loops.  Keep the loop
iterator the same, to ensure the list-post entries are comma
separated.

As a note, the parser is still not strictly following the RFC, but
appears to work well enough in the "real world" that I'm not going to
tweak it for now.