Remove cur parameter for pager recall and make-key ops.
For the recall operation, this can lead to a memory leak where the
recalled message overwrites the sctx saved data.

Arguably instead, the send_message setup could be more picky about
saving cur, or recall-message could be careful to not overwrite the
message-id.

The context parameter isn't actually used for mail-key, and the
curs_main.c doesn't pass it.  But I can't bring myself to make that
change for a stable fix.