Disable relative expansion for signature and source pipes.
Relative expansion shouldn't be performed when those values are to be
executed as a pipe for the output.

I believe this issue only affects $signature and the source command,
so for now I'm disabling relative expansion for those two cases.  If
there is too much more breakage, I will consider backing out the
relative expansion and disabling the :cd command.

Thanks to Aaron Schrab for reporting this issue.