Block signals during mbox-append operation.
I first noticed this back in 2017, but no one replied to my query to
mutt-dev and I forgot about it too.

Thanks to Oswald Buddenhagen for the followup when he was going
through the mailing list archives, confirming the mistake.

I've also reviewed MUTT_APPEND and MUTT_NEWFOLDER uses to make sure
the context is properly closed (and the signals restored).