Delay $hostname setting until after the muttrc is evaluated.
Commit 5c5c34f2 made back in 1.6 changed Fqdn setting to use
gethostname() and getaddrinfo() to get the canonical domain.  This is
more accurate but can cause startup delays for systems where the DNS
resolution is not set up properly.

Because this occurred before muttrc reading, there was no workaround
except to "fix DNS".

Change Fqdn ($hostname) setting to occur after the muttrc and '-e'
argument processing occur.

This is a possible breaking change if users rely on $hostname inside
their muttrc, for example 'source "muttrc.$hostname"'.  The workaround
would be to put something like (depending on the system type): 'set
hostname = `hostname --fqdn`' in the muttrc above that invocation.

Also note that we still set Hostname (the internal variable) early,
because it is used in more places than Fqdn, such as tempfiles.