Separate special color setting from $allow_ansi colors.
The ColorDefs for MT_COLOR_BOLD and MT_COLOR_UNDERLINE should only be
applied for the special formatting.  They shouldn't apply when
$allow_ansi has a bold/underline sequence.

Change the exclusive or operator to a regular or.  Exclusive or
implies we are doing some kind of toggling, which we are not in this
case.  It just confuses things.

Don't make the ansi underline, reverse, and blink mutually exclusive.