Disable normalization after expand_path().
The algorithm used was incorrect, for '..' expansion with symlinks
involved.  Furthermore, mutt_pretty_mailbox() takes care of this for
us.

To be conservative, just before the release, I'm leaving the function
but converting it to a noop.  I'll pull the function out after the
release.

Thanks to Oswald Buddenhagen for pointing out the bug!