Remove buffer_normalize_fullpath().
This was supposed to normalize a path after relative path expansion.
However, as Oswald Buddenhagen pointed out just before the release,
its method of '..' expansion was incorrect with symlinks in play.
Furthermore, it turned out mutt_pretty_mailbox() deals with '.' and
'..'.

The function was turned into a noop before the release, and I'm now
pulling it out.