add dprintf(n, fmt, ...) debugging macro
dprintf(n, fmt, ...) is a more natural form of
dprint(n, (debugfile, fmt, ...)). It prints code location in the debug
file where possible.

I believe this approach, while common, was not previously used because
we were trying to be C90-compatible.  C90 doesn't provide variadic
macros.  Now we require C99, so it makes sense to add this pattern.