Add configure error if C compiler doesn't support C99.
We added the test, but did not check to ensure it was found.  Thanks
to Oswald Buddenhagen for pointing out the missing check.  He
suggested the AC_PROG_CC might be redundant, but the info page seems
to indicate AC_PROG_CC_C99 checks the properties of an already located
compiler, so I'm leaving AC_PROG_CC in.

Also, remove the $U check just below.  Commit 64b1460a removed the
call to AM_C_PROTOTYPES, but forgot to remove the $U check too.