Fix exact-address handling when addr->personal is set.
The exact-address compile-time option takes an exact copy of an
address when it is parsed, and prints that out when outputting the
address.  The idea is to preserve older "user@host (Name)" syntax.

Unfortunately, when code tries to "update" the personal/name field, it
needs to clear the exact-address copy for it to have any effect.

An object-oriented design encapsulating the setting would help prevent
this problem.  It might be desirable to create a C function instead,
but callers would have to remember it, and such a thing isn't common
in the Mutt codebase.

Another patch (I'm still debating applying) changes the address parser
to discard the exact-address copy when it exactly matches the
mailbox.  However, that still won't fix every case (and it makes the
parser even more difficult to follow.)

So this commit (to stable) takes the straightforward approach.

It fixes bugs in:

* Alias creation.  The "personal name" prompt was ignored.

* Autocrypt initialization from address setting with $realname.

* $pgp_getkeys_command handling.

* Query menu results.

* "unset $reverse_realname" handling.

* $from handling of $realname.

* Bounce Resent-From handling of $realname.