Add <skip-headers> to skip past message headers in pager.
This is based on <skip-quoted>, and is bound to 'H' by default.

I've added an error message in case there is no text past the headers,
just to cover the logical case.  I don't think Mutt allows that to
happen for text passed to the pager, but just in case.