Fix REPLY_TO environment variable handling.
Commit 4e153adf changed this code to reuse the function buffer
variable, but forgot to rewind the buffer for parsing in
parse_my_hdr().

Additionally commit e5a32a61 removed an extra "null termination"
mutt_buffer_addch() at the end of mutt_extract_token().  This caused a
NULL value to be passed to the strpbrk() in parse_my_hdr(), causing a
segv.  Change to use a buffer pool token parameter instead.

I actually think, like with the previous IMAP mailbox handling, this
method of adding a my_hdr is dangerous.  I'll look into refactoring it
in master instead.

Thanks to Paul Nevai for reporting the problem and tracking down the
backtrace.